@ReneWerner87 2025-02-22
can you share benchmarks ? is it a speed improvement