@ksw2000 2025-02-22
> can you share benchmarks ? is it a speed improvement

The main purpose is to improve code readability because the original approach stored an integer in the map without ever reading it, which is quite unreasonable. I think this PR has negligible benefit in terms of performance.

```go
func Benchmark_Utils_UniqueRouteStack(b *testing.B) {
	for i := 0; i < b.N; i++ {

		route1 := &Route{}
		route2 := &Route{}
		route3 := &Route{}
		require.Equal(
			b,
			[]*Route{
				route1,
				route2,
				route3,
			},
			uniqueRouteStack([]*Route{
				route1,
				route1,
				route1,
				route2,
				route2,
				route2,
				route3,
				route3,
				route3,
				route1,
				route2,
				route3,
			}))
	}
}
```

```
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3
cpu: AMD EPYC 7763 64-Core Processor                
                          │  old2.txt   │           new2.txt            │
                          │   sec/op    │   sec/op     vs base          │
_Utils_UniqueRouteStack-4   1.364µ ± 1%   1.354µ ± 1%  ~ (p=0.133 n=50)

                          │  old2.txt  │            new2.txt            │
                          │    B/op    │    B/op     vs base            │
_Utils_UniqueRouteStack-4   704.0 ± 0%   704.0 ± 0%  ~ (p=1.000 n=50) ¹
¹ all samples are equal

                          │  old2.txt  │            new2.txt            │
                          │ allocs/op  │ allocs/op   vs base            │
_Utils_UniqueRouteStack-4   9.000 ± 0%   9.000 ± 0%  ~ (p=1.000 n=50) ¹
¹ all samples are equal
```