@ksw2000 2025-02-22
This benchmark tests the performance of inserting 16 different `Route`, with each `Route` being repeated twice, resulting in a total of 32 `Route` pointers being inserted into the `uniqueRouteStack` function.

```go
func Benchmark_Utils_UniqueRouteStack(b *testing.B) {
	routers := make([]Route, 16)
	for i := 0; i < b.N; i++ {
		uniqueRouteStack([]*Route{
			&routers[0],
			&routers[0],
			&routers[1],
			&routers[1],
			&routers[2],
			&routers[2],
			&routers[3],
			&routers[3],
			&routers[4],
			&routers[4],
			&routers[5],
			&routers[5],
			&routers[6],
			&routers[6],
			&routers[7],
			&routers[7],
			&routers[8],
			&routers[8],
			&routers[9],
			&routers[9],
			&routers[10],
			&routers[10],
			&routers[11],
			&routers[11],
			&routers[12],
			&routers[12],
			&routers[13],
			&routers[13],
			&routers[14],
			&routers[14],
			&routers[15],
			&routers[15],
		})
	}
}
```

Benchmark Results:
1. Time per operation: The new implementation is **9%** faster
2. Memory usage: The new implementation reduces memory usage by **38.34%**

```
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3
cpu: AMD EPYC 7763 64-Core Processor                
                          │   old.txt   │          new.txt           │
                          │   sec/op    │   sec/op     vs base       │
_Utils_UniqueRouteStack-4   1.561µ ± 1%   1.420µ ± 1%  -9.00% (n=50)

                          │   old.txt   │          new.txt           │
                          │    B/op     │    B/op     vs base        │
_Utils_UniqueRouteStack-4   1218.0 ± 0%   751.0 ± 0%  -38.34% (n=50)

                          │  old.txt   │            new.txt             │
                          │ allocs/op  │ allocs/op   vs base            │
_Utils_UniqueRouteStack-4   7.000 ± 0%   7.000 ± 0%  ~ (p=1.000 n=50) ¹
¹ all samples are equal
```

> can you share benchmarks ? is it a speed improvement